refactor(GameAccessRequest.java): 필드 타입 변경
타입을 일치시키기 위해 아이디를 저장하는 필드 타입을 long으로 변경

BREKAING CHANGE: gameId 필드 타입을 int에서 long으로 변경
as-is:
public class GameAccessRequest {

    private int gameId;
    ...

    public GameAccessRequest() {

    }

    public GameAccessRequest(int gameId, String password) {
        this.gameId = gameId;
        ...
    }

    public int getGameId() {
        ...
    }

    ...
}

public class GameAccessRequest {

    private long gameId;
    ...

    public GameAccessRequest() {

    }

    public GameAccessRequest(long gameId, String password) {
        this.gameId = gameId;
        ...
    }

    public long getGameId() {
        ...
    }

    ...
}

+++ src/main/java/chess/controller/dto/request/GameAccessRequest.java
@@ -2,19 +2,19 @@ package chess.controller.dto.request;
 
 public class GameAccessRequest {
 
-    private int gameId;
+    private long gameId;
     private String password;
 
     public GameAccessRequest() {
 
     }
 
-    public GameAccessRequest(int gameId, String password) {
+    public GameAccessRequest(long gameId, String password) {
         this.gameId = gameId;
         this.password = password;
     }
 
-    public int getGameId() {
+    public long getGameId() {
         return gameId;
     }
